+++ server/weaverbird/steps/__init__.py
@@ -28,6 +28,7 @@ from .select import SelectStep
 from .sort import SortStep
 from .split import SplitStep
 from .statistics import StatisticsStep
+from .substring import SubstringStep
 from .text import TextStep
 from .todate import ToDateStep
 from .top import TopStep