+++ src/lib/translators/pandas.ts
@@ -153,4 +153,8 @@ export class PandasTranslator extends BaseTranslator {
   split(step: Readonly<S.SplitStep>) {
     return step;
   }
+
+  substring(step: Readonly<S.SubstringStep>) {
+    return step;
+  }
 }